[FIX] 방만들기 모달창 추상화 추가
